Ensure FwLinks for all versions
FwLink is the field which is used in the code for all driver versions. Hence, the comment at the top of the resource files says to "maintain FwLinks for all versions". Since setting up and keeping actual FwLinks is a pain for so many versions, I simple renamed the DirLink to FwLink as new versions kept getting added.
I've tried to edit the file to maintain that.